requirements.txt
fastapi
uvicorn
curl_cffi
beautifulsoup4
python-dotenv